Revert "Feature/1.7 text to speech"
