--- Game/View/TupleInt.java
@@ -1,11 +1,0 @@
-package Game.View;
-
-public class TupleInt {
-    public final int x;
-    public final int y;
-
-    public TupleInt(int x, int y) {
-        this.x = x;
-        this.y = y;
-    }
-}
